Refactor the helpers into separate traversal/renaming/rewriting helper files.
